+++ cognite/client/_api/files.py
@@ -26,6 +26,8 @@ class FilesAPI(APIClient):
         asset_ids: List[int] = None,
         root_asset_ids: List[int] = None,
         root_asset_external_ids: List[str] = None,
+        asset_subtree_ids: List[int] = None,
+        asset_subtree_external_ids: List[str] = None,
         source: str = None,
         created_time: Union[Dict[str, Any], TimestampRange] = None,
         last_updated_time: Union[Dict[str, Any], TimestampRange] = None,
@@ -48,11 +50,13 @@ class FilesAPI(APIClient):
             asset_ids (List[int]): Only include files that reference these specific asset IDs.
             root_asset_ids (List[int]): The IDs of the root assets that the related assets should be children of.
             root_asset_external_ids (List[str]): The external IDs of the root assets that the related assets should be children of.
+            asset_subtree_ids (List[int]): List of asset subtrees ids to filter on.
+            asset_subtree_external_ids (List[str]): List of asset subtrees external ids to filter on.
             source (str): The source of this event.
             source_created_time (Union[Dict[str, Any], TimestampRange]): Filter for files where the sourceCreatedTime field has been set and is within the specified range.
             source_modified_time (Union[Dict[str, Any], TimestampRange]): Filter for files where the sourceModifiedTime field has been set and is within the specified range.
-            created_time (Union[Dict[str, Any], TimestampRange]): Range between two timestamps
-            last_updated_time (Union[Dict[str, Any], TimestampRange]): Range between two timestamps
+            created_time (Union[Dict[str, int], TimestampRange]):  Range between two timestamps. Possible keys are `min` and `max`, with values given as time stamps in ms.
+            last_updated_time (Union[Dict[str, int], TimestampRange]):  Range between two timestamps. Possible keys are `min` and `max`, with values given as time stamps in ms.
             uploaded_time (Union[Dict[str, Any], TimestampRange]): Range between two timestamps
             external_id_prefix (str): External Id provided by client. Should be unique within the project.
             uploaded (bool): Whether or not the actual file is uploaded. This field is returned only by the API, it has no effect in a post body.
@@ -64,12 +68,16 @@ class FilesAPI(APIClient):
         """
         if root_asset_ids or root_asset_external_ids:
             root_asset_ids = self._process_ids(root_asset_ids, root_asset_external_ids, wrap_ids=True)
+        if asset_subtree_ids or asset_subtree_external_ids:
+            asset_subtree_ids = self._process_ids(asset_subtree_ids, asset_subtree_external_ids, wrap_ids=True)
+
         filter = FileMetadataFilter(
             name=name,
             mime_type=mime_type,
             metadata=metadata,
             asset_ids=asset_ids,
             root_asset_ids=root_asset_ids,
+            asset_subtree_ids=asset_subtree_ids,
             source=source,
             created_time=created_time,
             last_updated_time=last_updated_time,
@@ -191,6 +199,8 @@ class FilesAPI(APIClient):
         asset_ids: List[int] = None,
         root_asset_ids: List[int] = None,
         root_asset_external_ids: List[str] = None,
+        asset_subtree_ids: List[int] = None,
+        asset_subtree_external_ids: List[str] = None,
         source: str = None,
         created_time: Union[Dict[str, Any], TimestampRange] = None,
         last_updated_time: Union[Dict[str, Any], TimestampRange] = None,
@@ -210,9 +220,11 @@ class FilesAPI(APIClient):
             asset_ids (List[int]): Only include files that reference these specific asset IDs.
             root_asset_ids (List[int]): The IDs of the root assets that the related assets should be children of.
             root_asset_external_ids (List[str]): The external IDs of the root assets that the related assets should be children of.
+            asset_subtree_ids (List[int]): List of asset subtrees ids to filter on.
+            asset_subtree_external_ids (List[str]): List of asset subtrees external ids to filter on.
             source (str): The source of this event.
-            created_time (Union[Dict[str, Any], TimestampRange]): Range between two timestamps
-            last_updated_time (Union[Dict[str, Any], TimestampRange]): Range between two timestamps
+            created_time (Union[Dict[str, int], TimestampRange]):  Range between two timestamps. Possible keys are `min` and `max`, with values given as time stamps in ms.
+            last_updated_time (Union[Dict[str, int], TimestampRange]):  Range between two timestamps. Possible keys are `min` and `max`, with values given as time stamps in ms.
             uploaded_time (Union[Dict[str, Any], TimestampRange]): Range between two timestamps
             source_created_time (Union[Dict[str, Any], TimestampRange]): Filter for files where the sourceCreatedTime field has been set and is within the specified range.
             source_modified_time (Union[Dict[str, Any], TimestampRange]): Filter for files where the sourceModifiedTime field has been set and is within the specified range.
@@ -248,12 +260,16 @@ class FilesAPI(APIClient):
         """
         if root_asset_ids or root_asset_external_ids:
             root_asset_ids = self._process_ids(root_asset_ids, root_asset_external_ids, wrap_ids=True)
+        if asset_subtree_ids or asset_subtree_external_ids:
+            asset_subtree_ids = self._process_ids(asset_subtree_ids, asset_subtree_external_ids, wrap_ids=True)
+
         filter = FileMetadataFilter(
             name=name,
             mime_type=mime_type,
             metadata=metadata,
             asset_ids=asset_ids,
             root_asset_ids=root_asset_ids,
+            asset_subtree_ids=asset_subtree_ids,
             source=source,
             created_time=created_time,
             last_updated_time=last_updated_time,