+++ cognite/client/_api/sequences.py
@@ -4,7 +4,14 @@ from typing import *
 
 from cognite.client import utils
 from cognite.client._api_client import APIClient
-from cognite.client.data_classes import Sequence, SequenceData, SequenceFilter, SequenceList, SequenceUpdate
+from cognite.client.data_classes import (
+    Sequence,
+    SequenceData,
+    SequenceDataList,
+    SequenceFilter,
+    SequenceList,
+    SequenceUpdate,
+)
 
 
 class SequencesAPI(APIClient):
@@ -23,6 +30,8 @@ class SequencesAPI(APIClient):
         metadata: Dict[str, str] = None,
         asset_ids: List[int] = None,
         root_asset_ids: List[int] = None,
+        asset_subtree_ids: List[int] = None,
+        asset_subtree_external_ids: List[str] = None,
         created_time: Dict[str, Any] = None,
         last_updated_time: Dict[str, Any] = None,
         limit: int = None,
@@ -38,19 +47,26 @@ class SequencesAPI(APIClient):
             metadata (Dict[str, Any]): Filter out sequences that do not match these metadata fields and values (case-sensitive). Format is {"key1":"value1","key2":"value2"}.
             asset_ids (List[int]): Filter out sequences that are not linked to any of these assets.
             root_asset_ids (List[int]): Filter out sequences not linked to assets with one of these assets as the root asset.
-            created_time (Dict[str, Any]): Filter out sequences with createdTime outside this range.
-            last_updated_time (Dict[str, Any]): Filter out sequences with lastUpdatedTime outside this range.
+            asset_subtree_ids (List[int]): List of asset subtrees ids to filter on.
+            asset_subtree_external_ids (List[str]): List of asset subtrees external ids to filter on.
+            created_time (Union[Dict[str, int], TimestampRange]):  Range between two timestamps. Possible keys are `min` and `max`, with values given as time stamps in ms.
+            last_updated_time (Union[Dict[str, int], TimestampRange]):  Range between two timestamps. Possible keys are `min` and `max`, with values given as time stamps in ms.
             limit (int, optional): Max number of sequences to return. Defaults to return all items.
 
         Yields:
             Union[Sequence, SequenceList]: yields Sequence one by one if chunk is not specified, else SequenceList objects.
         """
+
+        if asset_subtree_ids or asset_subtree_external_ids:
+            asset_subtree_ids = self._process_ids(asset_subtree_ids, asset_subtree_external_ids, wrap_ids=True)
+
         filter = SequenceFilter(
             name=name,
             metadata=metadata,
             external_id_prefix=external_id_prefix,
             asset_ids=asset_ids,
             root_asset_ids=root_asset_ids,
+            asset_subtree_ids=asset_subtree_ids,
             created_time=created_time,
             last_updated_time=last_updated_time,
         ).dump(camel_case=True)
@@ -130,6 +146,8 @@ class SequencesAPI(APIClient):
         metadata: Dict[str, str] = None,
         asset_ids: List[int] = None,
         root_asset_ids: List[int] = None,
+        asset_subtree_ids: List[int] = None,
+        asset_subtree_external_ids: List[str] = None,
         created_time: Dict[str, Any] = None,
         last_updated_time: Dict[str, Any] = None,
         limit: Optional[int] = 25,
@@ -144,8 +162,10 @@ class SequencesAPI(APIClient):
             metadata (Dict[str, Any]): Filter out sequences that do not match these metadata fields and values (case-sensitive). Format is {"key1":"value1","key2":"value2"}.
             asset_ids (List[int]): Filter out sequences that are not linked to any of these assets.
             root_asset_ids (List[int]): Filter out sequences not linked to assets with one of these assets as the root asset.
-            created_time (Dict[str, Any]): Filter out sequences with createdTime outside this range.
-            last_updated_time (Dict[str, Any]): Filter out sequences with lastUpdatedTime outside this range.
+            asset_subtree_ids (List[int]): List of asset subtrees ids to filter on.
+            asset_subtree_external_ids (List[str]): List of asset subtrees external ids to filter on.
+            created_time (Union[Dict[str, int], TimestampRange]):  Range between two timestamps. Possible keys are `min` and `max`, with values given as time stamps in ms.
+            last_updated_time (Union[Dict[str, int], TimestampRange]):  Range between two timestamps. Possible keys are `min` and `max`, with values given as time stamps in ms.
             limit (int, optional): Max number of sequences to return. Defaults to 25. Set to -1, float("inf") or None
                 to return all items.
 
@@ -174,12 +194,16 @@ class SequencesAPI(APIClient):
                 >>> for seq_list in c.sequences(chunk_size=2500):
                 ...     seq_list # do something with the sequences
         """
+        if asset_subtree_ids or asset_subtree_external_ids:
+            asset_subtree_ids = self._process_ids(asset_subtree_ids, asset_subtree_external_ids, wrap_ids=True)
+
         filter = SequenceFilter(
             name=name,
             metadata=metadata,
             external_id_prefix=external_id_prefix,
             asset_ids=asset_ids,
             root_asset_ids=root_asset_ids,
+            asset_subtree_ids=asset_subtree_ids,
             created_time=created_time,
             last_updated_time=last_updated_time,
         ).dump(camel_case=True)
@@ -499,10 +523,10 @@ class SequencesDataAPI(APIClient):
         start: int,
         end: Union[int, None],
         column_external_ids: Optional[List[str]] = None,
-        external_id: str = None,
-        id: int = None,
+        external_id: Union[str, List[str]] = None,
+        id: Union[int, List[int]] = None,
         limit: int = None,
-    ) -> SequenceData:
+    ) -> Union[SequenceData, SequenceDataList]:
         """`Retrieve data from a sequence <https://docs.cognite.com/api/v1/#operation/getSequenceData>`_
 
         Args:
@@ -512,7 +536,7 @@ class SequencesDataAPI(APIClient):
             column_external_ids (Optional[List[str]]): List of external id for the columns of the sequence. If 'None' is passed, all columns will be retrieved.
             id (int): Id of sequence.
             external_id (str): External id of sequence.
-            limit (int): Maximum number of rows to return.
+            limit (int): Maximum number of rows to return per sequence.
 
 
         Returns:
@@ -528,15 +552,29 @@ class SequencesDataAPI(APIClient):
                 >>> col = res.get_column(external_id='columnExtId') # ... get the array of values for a specific column,
                 >>> df = res.to_pandas() # ... or convert the result to a dataframe
         """
-        utils._auxiliary.assert_exactly_one_of_id_or_external_id(id, external_id)
-        post_obj = self._process_ids(id, external_id, wrap_ids=True)[0]
-        post_obj.update(self._process_columns(column_external_ids=column_external_ids))
-        post_obj.update({"start": start, "end": end, "limit": limit})
-        seqdata = []
-        columns = []
-        for data, columns in self._fetch_data(post_obj):
-            seqdata.extend(data)
-        return SequenceData(id=id, external_id=external_id, rows=seqdata, columns=columns)
+        post_objs = self._process_ids(id, external_id, wrap_ids=True)
+
+        def _fetch_sequence(post_obj):
+            post_obj.update(self._process_columns(column_external_ids=column_external_ids))
+            post_obj.update({"start": start, "end": end, "limit": limit})
+            seqdata = []
+            columns = []
+            for data, columns in self._fetch_data(post_obj):
+                seqdata.extend(data)
+            return SequenceData(
+                id=post_obj.get("id"), external_id=post_obj.get("externalId"), rows=seqdata, columns=columns
+            )
+
+        tasks_summary = utils._concurrency.execute_tasks_concurrently(
+            _fetch_sequence, [(x,) for x in post_objs], max_workers=self._config.max_workers
+        )
+        if tasks_summary.exceptions:
+            raise tasks_summary.exceptions[0]
+        results = tasks_summary.joined_results()
+        if len(post_objs) == 1:
+            return results[0]
+        else:
+            return SequenceDataList(results)
 
     def retrieve_dataframe(
         self,
@@ -544,6 +582,7 @@ class SequencesDataAPI(APIClient):
         end: Union[int, None],
         column_external_ids: Optional[List[str]] = None,
         external_id: str = None,
+        column_names: str = None,
         id: int = None,
         limit: int = None,
     ):
@@ -556,7 +595,8 @@ class SequencesDataAPI(APIClient):
             column_external_ids (Optional[List[str]]): List of external id for the columns of the sequence.  If 'None' is passed, all columns will be retrieved.
             id (int): Id of sequence
             external_id (str): External id of sequence.
-            limit (int): Maximum number of rows to return.
+            column_names (str):  Which field(s) to use as column header. Can use "externalId", "id", "columnExternalId", "id|columnExternalId" or "externalId|columnExternalId". Default is "externalId|columnExternalId" for queries on more than one sequence, and "columnExternalId" for queries on a single sequence.
+            limit (int): Maximum number of rows to return per sequence.
 
         Returns:
              pandas.DataFrame
@@ -567,7 +607,13 @@ class SequencesDataAPI(APIClient):
                 >>> c = CogniteClient()
                 >>> df = c.sequences.data.retrieve_dataframe(id=0, start=0, end=None)
         """
-        return self.retrieve(start, end, column_external_ids, external_id, id, limit).to_pandas()
+        if isinstance(external_id, List) or isinstance(id, List) or (id is not None and external_id is not None):
+            column_names_default = "externalId|columnExternalId"
+        else:
+            column_names_default = "columnExternalId"
+        return self.retrieve(start, end, column_external_ids, external_id, id, limit).to_pandas(
+            column_names=column_names or column_names_default
+        )
 
     def _fetch_data(self, task) -> Generator[Tuple[List, List], None, None]:
         remaining_limit = task.get("limit")