+++ tests/tests_unit/test_api/test_sequences.py
@@ -3,7 +3,14 @@ import re
 import pytest
 
 from cognite.client import CogniteClient
-from cognite.client.data_classes import Sequence, SequenceData, SequenceFilter, SequenceList, SequenceUpdate
+from cognite.client.data_classes import (
+    Sequence,
+    SequenceData,
+    SequenceDataList,
+    SequenceFilter,
+    SequenceList,
+    SequenceUpdate,
+)
 from tests.utils import jsgz_load
 
 COGNITE_CLIENT = CogniteClient()
@@ -165,6 +172,26 @@ class TestSequences:
     def test_list(self, mock_seq_response):
         res = SEQ_API.list()
         assert mock_seq_response.calls[0].response.json()["items"] == res.dump(camel_case=True)
+
+    def test_list_with_filters(self, mock_seq_response):
+        res = SEQ_API.list(
+            metadata={"a": "b"},
+            last_updated_time={"min": 45},
+            created_time={"max": 123},
+            asset_ids=[1, 2],
+            root_asset_ids=[1231],
+            asset_subtree_ids=[1],
+            asset_subtree_external_ids=["a"],
+        )
+        assert mock_seq_response.calls[0].response.json()["items"] == res.dump(camel_case=True)
+        assert {
+            "metadata": {"a": "b"},
+            "assetIds": [1, 2],
+            "assetSubtreeIds": [{"id": 1}, {"externalId": "a"}],
+            "rootAssetIds": [1231],
+            "createdTime": {"max": 123},
+            "lastUpdatedTime": {"min": 45},
+        } == jsgz_load(mock_seq_response.calls[0].request.body)["filter"]
 
     def test_create_single(self, mock_seq_response):
         res = SEQ_API.create(Sequence(external_id="1", name="blabla", columns=[{}]))
@@ -349,6 +376,21 @@ class TestSequences:
         assert 0 == len(data)
         assert 2 == len(data.column_external_ids)
 
+    def test_retrieve_multi_data(self, mock_seq_response, mock_get_sequence_data_two_col_with_zero):
+        data = SEQ_API.data.retrieve(id=[1, 2], start=0, end=None)
+        assert isinstance(data, SequenceDataList)
+        assert 2 == len(data)
+        assert 2 == len(data[0].column_external_ids)
+        assert 2 == len(data[1].column_external_ids)
+
+    def test_retrieve_multi_data_mixed(self, mock_seq_response, mock_get_sequence_data_two_col_with_zero):
+        data = SEQ_API.data.retrieve(external_id=["a"], id=[1, 2], start=0, end=None)
+        assert isinstance(data, SequenceDataList)
+        assert 3 == len(data)
+        assert 2 == len(data[0].column_external_ids)
+        assert 2 == len(data[1].column_external_ids)
+        assert 2 == len(data[2].column_external_ids)
+
     def test_retrieve_by_id(self, mock_seq_response, mock_get_sequence_data):
         data = SEQ_API.data.retrieve(id=123, start=123, end=None)
         assert isinstance(data, SequenceData)
@@ -435,6 +477,28 @@ class TestSequencesPandasIntegration:
         data = SEQ_API.data.retrieve(external_id="foo", start=1000000, end=1100000)
         assert isinstance(data, SequenceData)
         assert ["col1", "col2"] == list(data.to_pandas().columns)
+
+    def test_retrieve_dataframe_multi(self, mock_seq_response, mock_get_sequence_data_two_col):
+        data = SEQ_API.data.retrieve(external_id="foo", id=[2], start=1000000, end=1100000)
+        assert isinstance(data, SequenceDataList)
+        assert ["col1", "col2", "col1", "col2"] == list(data.to_pandas(column_names="columnExternalId").columns)
+
+    def test_to_pandas_invalid(self, mock_seq_response, mock_get_sequence_data_two_col):
+        with pytest.raises(ValueError):
+            SEQ_API.data.retrieve_dataframe(
+                external_id="foo", id=[2], start=1000000, end=1000001, column_names="id~columnExternalId"
+            )
+        with pytest.raises(ValueError):
+            SEQ_API.data.retrieve_dataframe(
+                external_id="foo", id=[2], start=1000000, end=1000001, column_names="columnExternalIds"
+            )
+
+    def test_retrieve_dataframe_column_names(self, mock_seq_response, mock_get_sequence_data_two_col):
+        data = SEQ_API.data.retrieve(external_id=["eid", "eid"], start=1000000, end=1100000)
+        assert isinstance(data, SequenceDataList)
+        assert ["col1", "col2", "col1", "col2"] == list(data.to_pandas(column_names="columnExternalId").columns)
+        assert ["eid", "eid", "eid", "eid"] == list(data.to_pandas(column_names="externalId").columns)
+        assert ["eid|col1", "eid|col2", "eid|col1", "eid|col2"] == list(data.to_pandas().columns)
 
     def test_retrieve_dataframe_columns_mixed_with_zero(
         self, mock_seq_response, mock_get_sequence_data_two_col_with_zero